fix to allow float values for lat,lon

+++ CropCircles.py
@@ -3,7 +3,7 @@ import subprocess
 
 if __name__ == '__main__':
     #use 40, -88
-    lat = int(input("Input latitude coordinate:"))
-    lon = int(input("Input longitude coordinate:"))
+    lat = float(input("Input latitude coordinate:"))
+    lon = float(input("Input longitude coordinate:"))
     save_map(lat, lon)
     subprocess.run(['java', '-cp', '.', 'AddCropCircles', 'map.png'])
